all sn_path and roman_path removed
